Splits reports page from report content
