+++ frontend/app/src/components/profitloss/ProfitLossOverview.vue
@@ -20,13 +20,21 @@
         <tr>
           <td>{{ $t('profit_loss_overview.rows.loan_profit') }}</td>
           <td class="text-right">
-            <amount-display pnl :value="overview.loanProfit" />
+            <amount-display
+              pnl
+              :value="overview.loanProfit"
+              :loading="loading"
+            />
           </td>
         </tr>
         <tr>
           <td>{{ $t('profit_loss_overview.rows.defi_profit_loss') }}</td>
           <td class="text-right">
-            <amount-display pnl :value="overview.defiProfitLoss" />
+            <amount-display
+              pnl
+              :value="overview.defiProfitLoss"
+              :loading="loading"
+            />
           </td>
         </tr>
         <tr>
@@ -34,13 +42,21 @@
             {{ $t('profit_loss_overview.rows.margin_positions_pnl') }}
           </td>
           <td class="text-right">
-            <amount-display pnl :value="overview.marginPositionsProfitLoss" />
+            <amount-display
+              pnl
+              :value="overview.marginPositionsProfitLoss"
+              :loading="loading"
+            />
           </td>
         </tr>
         <tr>
           <td>{{ $t('profit_loss_overview.rows.settlement_losses') }}</td>
           <td class="text-right">
-            <amount-display pnl :value="overview.settlementLosses" />
+            <amount-display
+              pnl
+              :value="overview.settlementLosses"
+              :loading="loading"
+            />
           </td>
         </tr>
         <tr>
@@ -48,7 +64,10 @@
             {{ $t('profit_loss_overview.rows.ethereum_transaction_gas_costs') }}
           </td>
           <td class="text-right">
-            <amount-display :value="overview.ethereumTransactionGasCosts" />
+            <amount-display
+              :value="overview.ethereumTransactionGasCosts"
+              :loading="loading"
+            />
           </td>
         </tr>
         <tr>
@@ -56,7 +75,10 @@
             {{ $t('profit_loss_overview.rows.asset_movement_fees') }}
           </td>
           <td class="text-right">
-            <amount-display :value="overview.assetMovementFees" />
+            <amount-display
+              :value="overview.assetMovementFees"
+              :loading="loading"
+            />
           </td>
         </tr>
         <tr>
@@ -64,7 +86,11 @@
             {{ $t('profit_loss_overview.rows.ledger_actions_profit_loss') }}
           </td>
           <td class="text-right">
-            <amount-display pnl :value="overview.ledgerActionsProfitLoss" />
+            <amount-display
+              pnl
+              :value="overview.ledgerActionsProfitLoss"
+              :loading="loading"
+            />
           </td>
         </tr>
         <tr>
@@ -72,7 +98,11 @@
             {{ $t('profit_loss_overview.rows.general_trade_profit_loss') }}
           </td>
           <td class="text-right">
-            <amount-display pnl :value="overview.generalTradeProfitLoss" />
+            <amount-display
+              pnl
+              :value="overview.generalTradeProfitLoss"
+              :loading="loading"
+            />
           </td>
         </tr>
         <tr>
@@ -80,7 +110,11 @@
             {{ $t('profit_loss_overview.rows.taxable_trade_profit_loss') }}
           </td>
           <td class="text-right text-subtitle-1">
-            <amount-display pnl :value="overview.taxableTradeProfitLoss" />
+            <amount-display
+              pnl
+              :value="overview.taxableTradeProfitLoss"
+              :loading="loading"
+            />
           </td>
         </tr>
         <tr>
@@ -91,7 +125,11 @@
             {{ $t('profit_loss_overview.rows.total_profit_loss') }}
           </td>
           <td class="text-right text-subtitle-1">
-            <amount-display pnl :value="overview.totalProfitLoss" />
+            <amount-display
+              pnl
+              :value="overview.totalProfitLoss"
+              :loading="loading"
+            />
           </td>
         </tr>
         <tr>
@@ -99,7 +137,11 @@
             {{ $t('profit_loss_overview.rows.total_taxable_profit_loss') }}
           </td>
           <td class="text-right text-subtitle-1">
-            <amount-display pnl :value="overview.totalTaxableProfitLoss" />
+            <amount-display
+              pnl
+              :value="overview.totalTaxableProfitLoss"
+              :loading="loading"
+            />
           </td>
         </tr>
       </tbody>
@@ -123,6 +165,11 @@ export default defineComponent({
       type: Object as PropType<ProfitLossOverview>
     },
     flat: {
+      required: false,
+      type: Boolean,
+      default: false
+    },
+    loading: {
       required: false,
       type: Boolean,
       default: false
