+++ frontend/app/src/router/routes.ts
@@ -11,7 +11,8 @@ export const Routes = {
   STAKING_ADEX: '/staking/adex',
   STAKING_LIQUITY: '/staking/liquity',
   DEFI_AIRDROPS: '/defi/airdrops',
-  PROFIT_LOSS_REPORT: '/profit-loss-report',
+  PROFIT_LOSS_REPORTS: '/reports/',
+  PROFIT_LOSS_REPORT: `/report/:id`,
   HISTORY_LEDGER_ACTIONS: '/history/ledger-actions',
   HISTORY_GITCOIN: '/history/gitcoin',
   ASSETS: '/assets/:identifier',
